requirements.txt
astropy==2.0.7
atomicwrites==1.1.5
attrs==18.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.3
bz2file==0.98
certifi==2019.6.16
cffi==1.11.5
config==0.4.2
configparser==3.5.0
cx-Oracle==7.1.2
cycler==0.10.0
Cython==0.28.3
decorator==4.3.0
dill==0.2.9
entrypoints==0.2.3
enum34==1.1.6
esutil==0.6.2.5
fitsio==0.9.11
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
h5py==2.8.0
healpy==1.12.0
html5lib==1.0.1
ipykernel==4.8.2
ipympl==0.3.1
ipython==5.7.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
linecache2==1.0.0
llvmlite==0.28.0
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
mock==2.0.0
more-itertools==4.2.0
mpi4py==2.0.0
multiprocess==0.70.7
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numba==0.43.1+0.g8dabe7abe.dirty
numexpr==2.6.9
numpy==1.14.5
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.2
pathos==0.2.3
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.6.0
pox==0.2.5
ppft==1.6.4.9
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.3
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.6.2
pytest-runner==4.2
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scandir==1.7
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
skymapper==0.1
subprocess32==3.5.2
tables==3.5.1
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traceback2==1.4.0
traitlets==4.3.2
unittest2==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
